requirements.txt
pandas
numpy
pyyaml
pyarrow
matplotlib
sentence-transformers
tqdm
hdbscan
httpx
jsonschema
python-dotenv
